relative paths -> absolute paths
