+++ src/MLComponents/CompoOptions/EDA/ColumnList.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
-import { targetURL, MLFUNC_URL, MLFUNC_SUFFIX_DF, URLS_PREPROCESS, httpConfig } from "./networkConfigs";
-import { AppContext } from "../../App";
-import { showDataResult } from "./util";
+import { targetURL, MLFUNC_URL, MLFUNC_SUFFIX_DF, URLS_PREPROCESS, httpConfig } from "MLComponents/CompoOptions/networkConfigs";
+import { AppContext } from "App";
+import { showDataResult } from "MLComponents/CompoOptions/util";
 
 function ColumnList({ formId, resultId }) {
   const { dfd, storage } = useContext(AppContext);
