+++ src/MLComponents/CompoOptions/EDA/Describe.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useContext } from "react";
-import { targetURL, MLFUNC_URL, MLFUNC_SUFFIX_DF, URLS_PREPROCESS, httpConfig } from "./networkConfigs";
-import { AppContext } from "../../App";
-import { inputStyle } from "../componentStyle";
-import { showDataResult } from "./util";
-import Switch from "./CompoPiece/Switch";
+import { targetURL, MLFUNC_URL, MLFUNC_SUFFIX_DF, URLS_PREPROCESS, httpConfig } from "MLComponents/CompoOptions/networkConfigs";
+import { AppContext } from "App";
+import { inputStyle } from "MLComponents/componentStyle";
+import { showDataResult } from "MLComponents/CompoOptions/util";
+import Switch from "MLComponents/CompoOptions/CompoPiece/Switch";
 
 function Describe({ formId, resultId }) {
   const [percentiles, setPercentiles] = useState(""); // 확인할 퍼센트 수치들
