+++ src/MLComponents/CompoOptions/EDA/Dtype.jsx
@@ -1,35 +1,26 @@
-import React, { useState, useContext } from "react";
-import { targetURL, MLFUNC_URL, MLFUNC_SUFFIX_DF, URLS_PREPROCESS, httpConfig } from "./networkConfigs";
-import { AppContext } from "../../App";
-import { inputStyle } from "../componentStyle";
-import { funcResultConfig, funcResultLayout } from "./funcResultConfigs";
-import { jsonToFile } from "./util";
+import React, { useContext } from "react";
+import { targetURL, MLFUNC_URL, MLFUNC_SUFFIX_DF, URLS_PREPROCESS, httpConfig } from "MLComponents/CompoOptions/networkConfigs";
+import { AppContext } from "App";
+import { funcResultConfig, funcResultLayout } from "MLComponents/CompoOptions/funcResultConfigs";
+import { jsonToFile } from "MLComponents/CompoOptions/util";
 
-function IsNa({ formId, resultId }) {
-  const [isSum, setIsSum] = useState();
+function Dtype({ formId, resultId }) {
   const { dfd, storage } = useContext(AppContext);
-
-  // 결측치 개수 합계 표시 여부 상태 값 저장
-  const handleChange = (event) => {
-    // console.log(event.target.value);
-    setIsSum(event.target.value);
-  };
 
   // 백앤드로 데이터 전송
   const handleSubmit = async (event) => {
     event.preventDefault(); // 실행 버튼 눌러도 페이지 새로고침 안 되도록 하는 것
 
     // 백앤드 전송을 위한 설정
-    const params = { sum: isSum }; // 입력해야 할 파라미터 설정
     // 백앤드 API URL에 파라미터 추가
-    const targetUrl = targetURL(MLFUNC_URL.concat(MLFUNC_SUFFIX_DF, URLS_PREPROCESS.IsNa), params);
+    const targetUrl = targetURL(MLFUNC_URL.concat(MLFUNC_SUFFIX_DF, URLS_PREPROCESS.Dtype));
     const df = storage.getItem("df"); // 기존에 스토리지에 저장되어 있던 데이터프레임(JSON) 가져오기
 
     // 데이터 전송 후 받아온 데이터프레임을 사용자에게 보여주기 위한 코드
     await fetch(targetUrl, httpConfig(JSON.stringify(df)))
       .then((response) => response.json())
       .then((data) => {
-        // console.log(data);
+        // document.getElementById(resultId).innerText = data;
         // JSON 데이터프레임 문자열을 담은 파일을 읽어서 데이터프레임으로 만든 후 보여주기
         dfd
           .readJSON(jsonToFile(data))
@@ -45,15 +36,9 @@ function IsNa({ formId, resultId }) {
 
   return (
     <form id={formId} onSubmit={handleSubmit}>
-      <label>
-        결측치 확인 방식
-        <select className={inputStyle} onChange={handleChange}>
-          <option value="false">데이터프레임</option>
-          <option value="true">컬럼별 결측치 수</option>
-        </select>
-      </label>
+      <span>옵션 없음</span>
     </form>
   );
 }
 
-export default React.memo(IsNa);
+export default React.memo(Dtype);
