+++ src/MLComponents/CompoOptions/EDA/index.js
@@ -1,0 +1,12 @@
+import ColumnList from "./ColumnList";
+import Corr from "./Corr";
+import Describe from "./Describe";
+import Dtype from "./Dtype";
+import Groupby from "./Groupby";
+import Head from "./Head";
+import IsNa from "./IsNa";
+import Shape from "./Shape";
+import Tail from "./Tail";
+import Unique from "./Unique";
+
+export { ColumnList, Corr, Describe, Dtype, Groupby, Head, IsNa, Shape, Tail, Unique };
