+++ src/MLComponents/CompoOptions/Preprocess/index.js
@@ -1,0 +1,23 @@
+import ColConditionDf from "./ColConditionDf";
+import Drop from "./Drop";
+import DropNa from "./DropNa";
+import ILocDf from "./ILocDf";
+import LocDf from "./LocDf";
+import RenameCol from "./RenameCol";
+import SortValue from "./SortValue";
+import Transpose from "./Transpose";
+// import SetColumn from "./SetColumn";
+// import MergeDf from "./MergeDf";
+// import ConcatDf from "./ConcatDf";
+
+export {
+  ColConditionDf,
+  Drop,
+  DropNa,
+  ILocDf,
+  LocDf,
+  RenameCol,
+  SortValue,
+  Transpose,
+  // MergeDf, ConcatDf, SetColumn,
+};
