+++ src/MLComponents/CompoOptions/index.js
@@ -1,55 +1,25 @@
 import DataUpload from "./DataUpload";
-
-// https://github.com/AI-Play/ML-Funcs/pull/43
-import Head from "./Head";
-import Tail from "./Tail";
-import Shape from "./Shape";
-import Dtype from "./Dtype";
-import Describe from "./Describe";
-import ColumnList from "./ColumnList";
-import Unique from "./Unique";
-import IsNa from "./IsNa";
-import Corr from "./Corr";
-import Transpose from "./Transpose";
-
-// https://github.com/AI-Play/ML-Funcs/pull/44
-import Groupby from "./Groupby";
-import Drop from "./Drop";
-import DropNa from "./DropNa";
-import RenameCol from "./RenameCol";
-import SortValue from "./SortValue";
-
-// https://github.com/AI-Play/ML-Funcs/pull/46
-// import MergeDf from "./MergeDf";
-// import ConcatDf from "./ConcatDf";
-
-// https://github.com/AI-Play/ML-Funcs/pull/52
-import LocDf from "./LocDf";
-import ILocDf from "./ILocDf";
-import ColConditionDf from "./ColConditionDf";
+import { ColumnList, Corr, Describe, Dtype, Groupby, Head, IsNa, Shape, Tail, Unique } from "./EDA";
+import { ColConditionDf, Drop, DropNa, ILocDf, LocDf, RenameCol, SortValue, Transpose } from "./Preprocess";
 
 export {
   DataUpload,
-  Head,
-  Tail,
-  Shape,
-  Dtype,
   ColumnList,
-  Unique,
-  IsNa,
   Corr,
   Describe,
-  ColConditionDf,
-  LocDf,
-  ILocDf,
-  Transpose,
+  Dtype,
   Groupby,
+  Head,
+  IsNa,
+  Shape,
+  Tail,
+  Unique,
+  ColConditionDf,
   Drop,
   DropNa,
+  ILocDf,
+  LocDf,
   RenameCol,
   SortValue,
-  // , , , , , ,
-  // , , , , ,
-  // MergeDf, ConcatDf,
-  // , ,
+  Transpose,
 };
