+++ tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  content: ["./src/**/*.{js,jsx,ts,tsx}", "./src/**/*.{html}", "./node_modules/tw-elements/dist/js/**/*.js"],
+  content: ["./src/**/*.{js,jsx,ts,tsx}", "./node_modules/tw-elements/dist/js/**/*.js"],
   theme: {
     extend: {},
   },
